# backend/docs/swagger.yaml
basePath: /
definitions:
...
  codepulse.TxBuildReq:
    properties:
      action:
        type: string
      params:
        additionalProperties: {}
        type: object
      wallet:
        type: string
    required:
    - action
    - wallet
    type: object
...
  handlers.ErrorJSON:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: 'REST API: health, chain status, counter contract calls, bank ledger
    (PostgreSQL indexer + optional The Graph subgraph).'
  title: go-chain API
  version: 0.1.0
paths:
...
  /api/code-pulse/proposals:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by organizer address
        in: query
        name: organizer
        type: string
      - description: Filter by round_review_state
        in: query
        name: review_state
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: page_size
        type: integer
      - description: 'Sort: submitted_at_desc (default), submitted_at_asc'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List proposals
      tags:
      - code-pulse
  /api/code-pulse/proposals/{proposalId}:
    get:
      parameters:
      - description: Proposal ID
        in: path
        name: proposalId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorJSON'
      summary: Proposal detail
      tags:
      - code-pulse
  /api/code-pulse/proposals/{proposalId}/timeline:
    get:
      parameters:
      - description: Proposal ID
        in: path
        name: proposalId
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Proposal timeline
      tags:
      - code-pulse
...
  /api/code-pulse/tx/{attemptId}:
    get:
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorJSON'
      summary: Get transaction attempt
      tags:
      - code-pulse
  /api/code-pulse/tx/build:
    post:
      consumes:
      - application/json
      parameters:
      - description: Action + wallet + params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codepulse.TxBuildReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Build transaction calldata
      tags:
      - code-pulse
  /api/code-pulse/tx/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Action + params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codepulse.TxBuildReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Submit transaction (server-signed)
      tags:
      - code-pulse
...
swagger: "2.0"